fix bug in approvals when transferring owner transfers tokens.
Prior to this commit, when the owner transferred a token to an address that had not owned SBT tokens before, that new token would not have an approval for the owner on it, so the owner would no longer be able to transfer it (and in fact, no one could).

Now, we ensure any address who receives an SBT will also give approval to the owner for that SBT to be transferred.